test: ENV_MUTEX/EnvGuard を test_env.rs に集約し crate 全体で共有
Codex 指摘対応: モジュールごとに別々の static ENV_MUTEX を持つと
crate 全体では相互排他できない。

- src/test_env.rs を新規作成し ENV_MUTEX (tokio::sync::Mutex) と
  EnvGuard (Drop ベースの自動復元) を一箇所に定義
- lib.rs と main.rs に #[cfg(test)] mod test_env; を追加
  → lib テスト binary / bin テスト binary それぞれで同一モジュールを共有
- config.rs と state.rs のテストは crate::test_env から import
